Fixed config bug that meant the game failed to load due to default value of an int being a float, Fixed packet data bug in particle accelerator tile misread a double as a float, leading to an incorrect value
The tile entity patch fixed the particle accelerator GUI bug where the particle velocity was incorrectly displayed

+++ src/main/java/org/halvors/nuclearphysics/common/tile/particle/TileParticleAccelerator.java
@@ -167,6 +167,8 @@ public class TileParticleAccelerator extends TileInventoryMachine implements IEl
                         entityParticle = null;
                     }
 
+
+
                     // Plays sound of particle accelerating past the speed based on total velocity at the time of anti-matter creation.
                     if (entityParticle != null) {
                         world.playSound(null, pos, ModSoundEvents.ACCELERATOR, SoundCategory.BLOCKS, 1.5F, (float) (0.6 + (0.4 * (entityParticle.getVelocity()) / ANTIMATTER_CREATION_SPEED)));
@@ -201,7 +203,7 @@ public class TileParticleAccelerator extends TileInventoryMachine implements IEl
         if (world.isRemote) {
             totalEnergyConsumed = dataStream.readInt();
             antimatterCount = dataStream.readInt();
-            velocity = dataStream.readFloat();
+            velocity = dataStream.readDouble();
         }
     }
